code structure change
