+++ src/propagation/main.cxx
@@ -1,6 +1,7 @@
 #include "PropagationAPI.h"
 #include "CommandLineHelper.h"
 #include "PropagationParameters.hxx"
+#include "PropagationInputBuilder.h"
 #include "PropagationIO.h"
 #include <iostream>
 #include <memory>
@@ -21,28 +22,28 @@ int usage()
   printf("usage: \n");
   printf("  greedy_propagation <options> <greedy_options> \n");
   printf("options for propagation: \n");
-  printf("  -spi <img4d.nii>              : 4D image that is the base of the segmentation \n");
-  printf("  -sps <seg.nii>                : 3D segmentation for the reference time point of the 4D base image. \n");
+  printf("  -i <img4d.nii>                : 4D image that is the base of the segmentation \n");
+  printf("  -sr3 <seg.nii>                : 3D segmentation for the reference time point of the 4D base image. \n");
   printf("                                  Only one reference segmentation input is allowed per run.\n");
-  printf("                                  This option will override all previous specified -sps or -sps-4d. \n");
-  printf("  -sps-4d <seg4d.nii>           : An alternative way to provide the reference segmentation. \n");
+  printf("                                  This option will override all previous specified -sr or -sr4. \n");
+  printf("  -sr4 <seg4d.nii>              : 4D segmentation providing the segmentation for the reference time point. \n");
   printf("                                  Only segementation from the reference time point will be used. \n");
-  printf("                                  This option will override all previous specified -sps or -sps-4d. \n");
-  printf("  -spo <outdir>                 : Output directory to store propagated segmentations and meshes\n");
-  printf("  -sps-op <pattern>             : Segmentation output filename pattern \n");
+  printf("                                  This option will override all previous specified -sr3 or -sr4. \n");
+  printf("  -tpr <timepoint>              : The reference time point of the given segmentation image \n");
+  printf("  -tpt <target tp str>          : A comma separated string of target time points for the propagation \n");
+  printf("  -o <outdir>                   : Output directory to store propagated segmentations and meshes\n");
+  printf("  -sr-op <pattern>              : (optional) Segmentation output filename pattern \n");
   printf("                                  The output filenames can be configured using a c-style pattern string, \n");
   printf("                                  with a timepoint number embedded.\n");
   printf("                                  For example: \"Seg_%%02d_resliced.nii.gz\" will generate \"Seg_05_resliced.nii.gz\" \n");
   printf("                                  for timepoint 5. \n");
+  printf("  -sr-mop <pattern>             : (optional) Set the filename pattern for the mesh generated from the provided reference segmentation image. \n");
   printf("                                  Filename pattern without %%format will have time point appended\n");
-  printf("  -sp-interp-spec <spec> <size> : Set the segmentation interpolation spec with following options: \n");
+  printf("  -int <spec> <size>            : (optional) Set the segmentation interpolation spec with following options: \n");
   printf("                                  NN; LINEAR; LABEL <size>\n");
-  printf("  -sps-mop <pattern>            : Set the filename pattern for the mesh generated from the provided reference segmentation image. \n");
-  printf("  -spm <mesh.vtk> <pattern>     : Add an extra segmentation mesh for the referenfe tp. Repeat to add multiple meshes. \n");
-  printf("  -spr <timepoint>              : The reference time point of the given segmentation image \n");
-  printf("  -spt <target tp str>          : A comma separated string of target time points for the propagation \n");
-  printf("  -sp-debug <outdir>            : Enable debugging mode for propagation: Dump intermediary files to outdir \n");
-  printf("  -sp-verbose <value>           : Set propagation verbosity level (0: none, 1: default, 2: verbose) \n");
+  printf("  -emr <mesh.vtk> <pattern>     : (optional) Add an extra segmentation mesh for the referenfe tp. Repeat to add multiple meshes. \n");
+  printf("  -debug <outdir>               : (optional) Enable debugging mode for propagation: Dump intermediary files to outdir \n");
+  printf("  -verbose <value>              : (optional) Set propagation verbosity level (0: none, 1: default, 2: verbose) \n");
   printf("main greedy options accepted: \n");
   printf("  ");
   for (auto cit = greedy_cmd.crbegin(); cit != greedy_cmd.crend(); ++cit)
@@ -50,6 +51,7 @@ int usage()
   printf("\n");
   return -1;
 }
+
 
 void parse_command_line(int argc, char *argv[],
                         PropagationParameters &pParam, GreedyParameters &gParam)
@@ -59,33 +61,32 @@ void parse_command_line(int argc, char *argv[],
 
   while (cl.read_command(arg))
     {
-    if (arg == "-spi")
+    if (arg == "-i")
       {
       pParam.fn_img4d = cl.read_existing_filename();
       }
-    else if (arg == "-spo")
+    else if (arg == "-o")
       {
       pParam.outdir = cl.read_output_dir();
       }
-    else if (arg == "-sps")
+    else if (arg == "-sr3")
       {
       pParam.fn_seg3d = cl.read_existing_filename();
-      pParam.use4DSegInput = false;
       }
-    else if (arg == "-sps-4d")
+    else if (arg == "-sr4")
       {
       pParam.fn_seg4d = cl.read_existing_filename();
       pParam.use4DSegInput = true;
       }
-    else if (arg == "-sps-op")
+    else if (arg == "-sr-op")
       {
       pParam.fnsegout_pattern = cl.read_string();
       }
-    else if (arg == "-sps-mop")
+    else if (arg == "-sr-mop")
       {
       pParam.fnmeshout_pattern = cl.read_string();
       }
-    else if (arg == "-spm")
+    else if (arg == "-emr")
       {
       MeshSpec meshspec;
       meshspec.cached = false; // mesh is from a file
@@ -93,11 +94,11 @@ void parse_command_line(int argc, char *argv[],
       meshspec.fnout_pattern = cl.read_string();
       pParam.extra_mesh_list.push_back(meshspec);
       }
-    else if (arg == "-spr")
+    else if (arg == "-tpr")
       {
       pParam.refTP = cl.read_integer();
       }
-    else if (arg == "-spt")
+    else if (arg == "-tpt")
       {
       std::vector<int> result = cl.read_int_vector(',');
       std::set<int> unique(result.begin(), result.end()); // remove duplicates
@@ -111,7 +112,7 @@ void parse_command_line(int argc, char *argv[],
         pParam.targetTPs.push_back(n);
         }
       }
-    else if (arg == "-sp-interp-spec")
+    else if (arg == "-int")
       {
       // propagation mode: read sigmas for label reslicing interpolation mode
       std::string mode = cl.read_string();
@@ -134,12 +135,12 @@ void parse_command_line(int argc, char *argv[],
         std::cerr << "Propagation interpolation spec: Unknown interpolation mode" << std::endl;
         }
       }
-    else if (arg == "-sp-debug")
+    else if (arg == "-debug")
       {
       pParam.debug = true;
       pParam.debug_dir = cl.read_output_dir();
       }
-    else if (arg == "-sp-verbose")
+    else if (arg == "-verbose")
       {
       int level = cl.read_integer();
       if(level < 0 || level >= PropagationParameters::VERB_INVALID)
@@ -157,10 +158,14 @@ void parse_command_line(int argc, char *argv[],
 }
 
 template<typename TReal>
-int run (PropagationParameters &pParam, GreedyParameters &gParam)
+int run (const PropagationParameters &pParam, const GreedyParameters &gParam)
 {
   std::cout << "-- [Propagation] Run started" << std::endl;
-  auto pInput = PropagationInputBuilder<TReal>::BuildInputForCommandLineRun(pParam, gParam);
+
+  PropagationInputBuilder<TReal> builder;
+  builder.ConfigForCLI(pParam, gParam);
+  auto pInput = builder.BuildPropagationInput();
+  
   PropagationAPI<TReal> api(pInput);
   return api.Run();
 }
@@ -170,8 +175,8 @@ int main (int argc, char *argv[])
   if (argc < 2)
     return usage();
 
-  PropagationParameters pParam;
-  GreedyParameters gParam;
+  PropagationParameters pParam; 
+  GreedyParameters gParam; // for general greedy parameters
 
   try
     {
